added valid dependency in pom.xml

+++ src/main/java/com/magz/backend/common/repository/UserRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-    public interface UserRepository extends CrudRepository<User, Integer> {
+    public interface UserRepository extends CrudRepository<User, Long> {
         Optional<User> findByEmail(String email); // Use 'email' if that is the correct field for login
+
     }
